fix(designer): Dynamic content was incorrectly cached for swagger based operations
